Feat : For compiling, separate functions and sychoronize withe bash error message!
